Added link to the social contract.
The social contract is hosted on google docs. This will put a link so
that it will be possible to quickly edit. Maybe a better solution is to
download the document to this folder. This has however the problem that
it has to be updated. A pure git flow is not really that good when
everyone is editing such a small *text* document at the same time.