Remove max-height attribute from Footer main css.
I removed the attribute max-height attribute from the class .main-footer
in the file Footer.css because it was to small and broke the rendering
of the background on footer. Please comment on this if you hade an
important reason to use max-heigt.
This seemed to fix it for me but I am not sure if there was a reason for
including it in the first place.