Added a java replacement of jsonplaceholder.
This is written using Spring boot IO and psotgressql. At the moment it
just copies the functionality that json server provided earlier but it
is very possible to extend. At the moment there isn't any documentation
for the code but it is in the works. To get postgres to play nice you
have to run the `tables.sql` in PSQL this can be done with `\i
tables.sql`. There is also a CLEAN file which drops *ALL* tables so be
careful that it is run in the right database. I have also written
inserts to get basic data up and running. At the moment I have no data
for the bookings but that can be written too. We should create a
setup.sql file to get up and running quickly.
If something is unclear please do not hesitate to contact me. Write on
Slack or comment on this commit.